+++ R/filter.R
@@ -45,7 +45,12 @@ filter.RPolarsDataFrame <- function(.data, ..., .by = NULL) {
   mo <- attributes(.data)$maintain_grp_order
   is_grouped <- !is.null(grps)
 
-  polars_exprs <- translate_dots(.data, ..., env = rlang::current_env())
+  polars_exprs <- translate_dots(
+    .data,
+    ...,
+    env = rlang::current_env(),
+    caller = rlang::caller_env()
+  )
 
   if (is_grouped) {
     polars_exprs <- lapply(polars_exprs, \(x) x$over(grps))