+++ R/funs-date.R
@@ -122,6 +122,10 @@ pl_dnanoseconds <- function(x, ...) {
   pl$duration(nanoseconds = x)
 }
 
+pl_make_date <- function(year = 1970, month = 1, day = 1) {
+  pl$date(year = year, month = month, day = day)
+}
+
 # Periods --------------------------------------
 
 
@@ -203,12 +207,6 @@ pl_replace_time_zone <- function(x, ...) {
   x$dt$replace_time_zone()
 }
 
-pl_round <- function(x, ...) {
-  check_empty_dots(...)
-  x$dt$round()
-}
-
-
 pl_seconds <- function(x, ...) {
   check_empty_dots(...)
   x$dt$seconds()
@@ -222,12 +220,11 @@ pl_strftime <- function(x, ...) {
 pl_strptime <- function(string, format, tz = "", strict = TRUE, ...) {
   check_empty_dots(...)
   if (grepl("%(I|H|c|T|M|p|r|R|S|X|z)", format)) {
-    datatype = pl$Datetime("us", tz = tz)
+    dtype = pl$Datetime("us", tz = tz)
   } else {
-    datatype = pl$Date
+    dtype = pl$Date
   }
-  # TODO: polars 0.16.0: replace datatype by dtype
-  string$str$strptime(datatype = datatype, format = format, strict = strict)
+  string$str$strptime(dtype = dtype, format = format, strict = strict)
 }
 
 pl_timestamp <- function(x, ...) {