+++ tests/tinytest/test_drop_na_lazy.R
@@ -32,6 +32,11 @@ expect_dim(
   c(28, 11)
 )
 
+expect_dim(
+  drop_na(pl_tmp),
+  c(28, 11)
+)
+
 expect_error_lazy(
   drop_na(pl_tmp, foo),
   "don't exist"