+++ tests/tinytest/test_filter_lazy.R
@@ -233,5 +233,18 @@ expect_equal_lazy(
   NULL
 )
 
+# works with a local variable defined in a function
+
+foobar <- function(x) {
+  local_var <- "a"
+  x |> filter(chars == local_var)
+}
+
+test <- polars::pl$LazyFrame(chars = letters[1:3])
+
+expect_equal_lazy(
+  foobar(test),
+  data.frame(chars = "a")
+)
 
 Sys.setenv('TIDYPOLARS_TEST' = FALSE)